requirements.txt
Django==3.1.7
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
html5lib==1.1
mysqlclient==2.0.1
openpyxl==3.0.7
pdfkit==0.6.1
Pillow==8.1.2
pytz==2021.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
